requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.34.0
datasets>=2.14.0
sentencepiece>=0.1.99
accelerate>=0.23.0
bitsandbytes>=0.41.1
peft>=0.5.0
tqdm>=4.66.1
pillow>=10.0.0
wandb>=0.15.11
requests>=2.31.0
